more warnings fix

+++ unix/render.c
@@ -839,8 +839,8 @@ my_XRenderCompositeDoublePoly (Display		    *dpy,
     edges = (Edge *) malloc (npoints * sizeof (Edge) +
 			      ((npoints * npoints + 1) * sizeof (XTrapezoid)));
     if (!edges) {
-        warn("XRenderCompositeDoublePoly: not enough memory (%d vertices, %d bytes)\n",
-		npoints, npoints * sizeof (Edge) + ((npoints * npoints + 1) * sizeof (XTrapezoid)));
+        warn("XRenderCompositeDoublePoly: not enough memory (%d vertices, %ld bytes)\n",
+		npoints, (long unsigned int)(npoints * sizeof (Edge) + ((npoints * npoints + 1) * sizeof (XTrapezoid))));
 	return 0;
     }
     traps = (XTrapezoid *) (edges + npoints);
